requirements.txt
flask
gocomics.py
